stream-processing/connect/requirements.txt
avro==1.12.1
confluent-kafka==2.4.0
requests==2.31.0
fastavro==1.7.0
psycopg2-binary==2.9.11
click==8.3.1